Deprecated some older code, added new dependencies
- Instead of using WifiPlug.Api.Schema.Characteristics, use WifiPlug.DeviceFramework.Characteristic
- ScanResult<TEntity> now uses a int64 for total instead of int32, same with most results entities
- Moved to 1.3.0

+++ src/WifiPlug.Api/Entities/DeviceResultsEntity.cs
@@ -29,6 +29,6 @@ namespace WifiPlug.Api.Entities
         /// Gets or sets the total number of devices.
         /// </summary>
         [JsonProperty(PropertyName = "total_device_count")]
-        public int TotalDevices { get; set; }
+        public long TotalDevices { get; set; }
     }
 }
